janus_proxy/requirements.txt
websockets>=11.0
pyyaml>=6.0
